Made inventory tests (#38)
* Made inventory tests

Moved tests for the inventory to its own file away from test_avatar.py

* Fixed Documentation

Fixed documentation for avatar inventory tests.

* Documentation

Updated documentation again.

---------
